+++ crates/fraud_proofs/global_merkle_root/storage/src/structured_storage/messages.rs
@@ -1,0 +1,11 @@
+use crate::{
+    column::TableColumn,
+    merkle::MerkleizedTableColumn,
+};
+use fuel_core_storage::tables::Messages;
+
+impl MerkleizedTableColumn for Messages {
+    fn table_column() -> TableColumn {
+        TableColumn::Messages
+    }
+}